chore: move OwlBotLock into config-files.ts
I need to re-use OwlBotLock's definition in the database code, and I intend to define .OwlBot.yaml stuff here too.
